NFC: Move common GPU passes under `Common/GPU/` directory
This improves the codebase structure and avoids unnecessary GPU
pass dependencies for CPU targets.
