+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -123,7 +123,6 @@ iree_compiler_cc_library(
         "FoldTensorExtractOpPass.cpp",
         "ForOpCanonicalizationPass.cpp",
         "FuseTensorPadWithConsumer.cpp",
-        "GPUPipelining.cpp",
         "HoistStaticallyBoundAllocations.cpp",
         "IREEComprehensiveBufferizePass.cpp",
         "IREEExpandStridedMetadata.cpp",
@@ -179,7 +178,6 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:BufferizationTransforms",
         "@llvm-project//mlir:DialectUtils",
         "@llvm-project//mlir:FuncDialect",
-        "@llvm-project//mlir:GPUDialect",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:LLVMCommonConversion",
         "@llvm-project//mlir:LLVMDialect",
@@ -190,7 +188,6 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:MathTransforms",
         "@llvm-project//mlir:MemRefDialect",
         "@llvm-project//mlir:MemRefTransforms",
-        "@llvm-project//mlir:NVGPUDialect",
         "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:SCFDialect",
         "@llvm-project//mlir:SCFToControlFlow",
@@ -217,26 +214,14 @@ iree_compiler_cc_library(
         "DecomposeLinalgGeneric.cpp",
         "ExtractAddressComputation.cpp",
         "FoldAffineMinInDistributedLoops.cpp",
-        "GPUDistribute.cpp",
-        "GPUDistributeSharedMemoryCopy.cpp",
-        "GPUMultiBuffering.cpp",
-        "GPUPatterns.cpp",
-        "GPUReduceBankConflicts.cpp",
-        "GPUTensorAlloc.cpp",
-        "GPUTensorTile.cpp",
-        "GPUTileReduction.cpp",
-        "GPUVectorization.cpp",
         "LinalgOpInfo.cpp",
         "MemrefCopyToLinalg.cpp",
         "PadDynamicAlloc.cpp",
         "RemoveTrivialLoops.cpp",
         "UserConfig.cpp",
-        "VectorReductionToGPU.cpp",
-        "WorkGroupSwizzle.cpp",
     ],
     hdrs = [
         "ExtractAddressComputation.h",
-        "GPUPatterns.h",
         "LinalgOpInfo.h",
         "UserConfig.h",
     ],
@@ -269,9 +254,6 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:BufferizationTransforms",
         "@llvm-project//mlir:DialectUtils",
         "@llvm-project//mlir:FuncDialect",
-        "@llvm-project//mlir:GPUDialect",
-        "@llvm-project//mlir:GPUTransformOps",
-        "@llvm-project//mlir:GPUTransforms",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:LLVMCommonConversion",
         "@llvm-project//mlir:LLVMDialect",
@@ -281,7 +263,6 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:MathTransforms",
         "@llvm-project//mlir:MemRefDialect",
         "@llvm-project//mlir:MemRefTransforms",
-        "@llvm-project//mlir:NVGPUDialect",
         "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:SCFDialect",
         "@llvm-project//mlir:SCFToControlFlow",
