+++ compiler/src/iree/compiler/Codegen/Common/GPU/CMakeLists.txt
@@ -1,0 +1,68 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# compiler/src/iree/compiler/Codegen/Common/GPU/BUILD.bazel                    #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    CommonGPUPasses
+  HDRS
+    "CommonGPUPasses.h"
+    "GPUPatterns.h"
+  SRCS
+    "GPUDistribute.cpp"
+    "GPUDistributeSharedMemoryCopy.cpp"
+    "GPUMultiBuffering.cpp"
+    "GPUPatterns.cpp"
+    "GPUPipelining.cpp"
+    "GPUReduceBankConflicts.cpp"
+    "GPUTensorAlloc.cpp"
+    "GPUTensorTile.cpp"
+    "GPUTileReduction.cpp"
+    "GPUVectorization.cpp"
+    "VectorReductionToGPU.cpp"
+    "WorkGroupSwizzle.cpp"
+  DEPS
+    IREELinalgExtDialect
+    IREELinalgExtTransforms
+    LLVMSupport
+    MLIRAffineDialect
+    MLIRAffineTransforms
+    MLIRAffineUtils
+    MLIRArithDialect
+    MLIRBufferizationDialect
+    MLIRFuncDialect
+    MLIRGPUOps
+    MLIRGPUTransformOps
+    MLIRGPUTransforms
+    MLIRIR
+    MLIRLinalgDialect
+    MLIRLinalgTransforms
+    MLIRLinalgUtils
+    MLIRMemRefDialect
+    MLIRMemRefTransforms
+    MLIRNVGPUDialect
+    MLIRSCFDialect
+    MLIRSCFTransforms
+    MLIRSCFUtils
+    MLIRSideEffectInterfaces
+    MLIRTransforms
+    MLIRVectorDialect
+    MLIRVectorToSCF
+    MLIRVectorTransforms
+    iree::compiler::Codegen::Dialect::IREECodegenDialect
+    iree::compiler::Codegen::PassHeaders
+    iree::compiler::Codegen::Transforms
+    iree::compiler::Codegen::Utils
+    iree::compiler::Dialect::HAL::IR
+  PUBLIC
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
