+++ compiler/src/iree/compiler/Codegen/Common/GPU/test/BUILD.bazel
@@ -1,0 +1,41 @@
+# Copyright 2023 The IREE Authors
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+# Tests for common transforms.
+
+load("//build_tools/bazel:iree_lit_test.bzl", "iree_lit_test_suite")
+load("//build_tools/bazel:enforce_glob.bzl", "enforce_glob")
+
+package(
+    features = ["layering_check"],
+    licenses = ["notice"],  # Apache 2.0
+)
+
+iree_lit_test_suite(
+    name = "lit",
+    srcs = enforce_glob(
+        [
+            "gpu_distribute.mlir",
+            "gpu_distribute_shared_memory.mlir",
+            "gpu_pipeline.mlir",
+            "gpu_tensor_alloc.mlir",
+            "gpu_tensor_tile.mlir",
+            "gpu_vectorization.mlir",
+            "gpu_workgroup_swizzle.mlir",
+            "gpu_tile_reduction.mlir",
+            "reduce_bank_conflicts.mlir",
+        ],
+        include = ["*.mlir"],
+        exclude = [
+        ],
+    ),
+    cfg = "//compiler:lit.cfg.py",
+    tools = [
+        "//tools:iree-opt",
+        "@llvm-project//llvm:FileCheck",
+    ],
+)
+
