+++ compiler/src/iree/compiler/Codegen/Common/GPU/test/CMakeLists.txt
@@ -1,0 +1,31 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# compiler/src/iree/compiler/Codegen/Common/GPU/test/BUILD.bazel               #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_lit_test_suite(
+  NAME
+    lit
+  SRCS
+    "gpu_distribute.mlir"
+    "gpu_distribute_shared_memory.mlir"
+    "gpu_pipeline.mlir"
+    "gpu_tensor_alloc.mlir"
+    "gpu_tensor_tile.mlir"
+    "gpu_tile_reduction.mlir"
+    "gpu_vectorization.mlir"
+    "gpu_workgroup_swizzle.mlir"
+    "reduce_bank_conflicts.mlir"
+  TOOLS
+    FileCheck
+    iree-opt
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
