+++ compiler/src/iree/compiler/Codegen/Common/test/CMakeLists.txt
@@ -25,7 +25,6 @@ iree_lit_test_suite(
     "decompose_affine_ops.mlir"
     "decompose_linalg_generic.mlir"
     "decompose_pack_unpack_ops.mlir"
-    "distribute_gpu_shared_memory.mlir"
     "eliminate_empty_tensors.mlir"
     "erase_hal_descriptor_type.mlir"
     "extract_address_computation.mlir"
@@ -34,27 +33,19 @@ iree_lit_test_suite(
     "fold_affine_min_of_block_id.mlir"
     "fold_tensor_extract_op.mlir"
     "forop_canonicalization.mlir"
-    "gpu_distribute.mlir"
-    "gpu_pipeline.mlir"
-    "gpu_tensor_alloc.mlir"
-    "gpu_tensor_tile.mlir"
-    "gpu_vectorization.mlir"
     "hoist_statically_bound_allocations.mlir"
     "iree_comprehensive_bufferize.mlir"
     "iree_expand_strided_metadata.mlir"
     "lower_ukernel_to_calls.mlir"
     "pad_dynamic_alloc.mlir"
     "polynomial_approximation.mlir"
-    "reduce_bank_conflicts.mlir"
     "reductions.mlir"
     "rematerialize_parallel_ops.mlir"
     "remove_dead_allocs.mlir"
     "remove_trivial_loops.mlir"
     "repeated_matcher_use.mlir"
-    "swizzle_workgroup.mlir"
     "test_partitionable_loops_interface.mlir"
     "tile_and_distribute_to_workgroups.mlir"
-    "tile_reduction.mlir"
     "transform_buffer_opt.mlir"
     "transform_dialect_apply_pattern_op.mlir"
     "transform_ops_invalid.mlir"
